% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11qpdf-package.R
\docType{package}
\name{cpp11qpdf-package}
\alias{cpp11qpdf-package}
\title{Split, Combine and Compress PDF Files}
\description{
Bindings to 'cpp11qpdf':
A tool for content-preserving transformations of PDF files such  as split,
combine, and compress.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.pachadotdev.org/cpp11qpdf/ https://pachadotdev.r-universe.dev/cpp11qpdf}
  \item Report bugs at \url{https://github.com/pachadotdev/cpp11qpdf/issues}
}

}
\author{
\strong{Maintainer}: Mauricio Vargas Sepulveda \email{m.sepulveda@mail.utoronto.ca} (\href{https://orcid.org/0000-0003-1017-7574}{ORCID})

Authors:
\itemize{
  \item Jeroen Ooms \email{jeroen@berkeley.edu} (\href{https://orcid.org/0000-0002-4035-0289}{ORCID})
}

Other contributors:
\itemize{
  \item Ben Raymond [contributor]
  \item Jay Berkenbilt (Author of libcpp11qpdf) [copyright holder]
  \item Munk School of Global Affairs and Public Policy [funder]
}

}
