% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{cpp11qpdf}
\alias{cpp11qpdf}
\alias{pdf_split}
\alias{pdf_length}
\alias{pdf_subset}
\alias{pdf_combine}
\alias{pdf_compress}
\alias{pdf_overlay_stamp}
\alias{pdf_rotate_pages}
\title{PDF Functions}
\usage{
pdf_split(input, output = NULL, password = askpass::askpass())

pdf_length(input, password = askpass::askpass())

pdf_subset(input, pages = 1, output = NULL, password = askpass::askpass())

pdf_combine(input, output = NULL, password = askpass::askpass())

pdf_compress(
  input,
  output = NULL,
  linearize = FALSE,
  password = askpass::askpass()
)

pdf_overlay_stamp(input, stamp, output = NULL, password = askpass::askpass())

pdf_rotate_pages(
  input,
  pages,
  angle = 90,
  relative = FALSE,
  output = NULL,
  password = askpass::askpass()
)
}
\arguments{
\item{input}{path or url to the input pdf file}

\item{output}{base path of the output file(s)}

\item{password}{string with password to open pdf file}

\item{pages}{a vector with page numbers to rotate}

\item{linearize}{enable pdf linearization (streamable pdf)}

\item{stamp}{pdf file of which the first page is overlayed into each page of input}

\item{angle}{rotation angle in degrees (positive = clockwise)}

\item{relative}{if \code{TRUE}, pages are rotated relative to their current orientation. If \code{FALSE}, rotation is absolute (0 = portrait, 90 = landscape, rotated 90 degrees clockwise from portrait)}
}
\description{
PDF Functions
}
\examples{
\donttest{
# extract some pages
pdf_file <- system.file("examples", "sufganiyot.pdf", package = "cpp11qpdf")
pdf_subset(pdf_file, 1, pdf_file, password = "")
pdf_length(pdf_file, password = "")
}
}
