% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{pdf_compress}
\alias{pdf_compress}
\title{Compress a pdf file}
\usage{
pdf_compress(input, output = NULL, linearize = FALSE, password = "")
}
\arguments{
\item{input}{path or url to the input pdf file}

\item{output}{base path of the output file(s)}

\item{linearize}{enable pdf linearization (streamable pdf)}

\item{password}{string with password to open pdf file}
}
\value{
a character vector with the path of the compressed pdf file
}
\description{
Compress a pdf file
}
\examples{
pdf_file <- system.file("examples", "sufganiyot.pdf", package = "cpp11qpdf")
fout <- tempfile()
pdf_compress(pdf_file, fout, TRUE, "")
}
