% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{pdf_length}
\alias{pdf_length}
\title{Get the number of pages in a pdf file}
\usage{
pdf_length(input, password = "")
}
\arguments{
\item{input}{path or url to the input pdf file}

\item{password}{string with password to open pdf file}
}
\value{
an integer value with the number of pages in the pdf file
}
\description{
Get the number of pages in a pdf file
}
\examples{
pdf_file <- system.file("examples", "sufganiyot.pdf", package = "cpp11qpdf")
pdf_length(pdf_file, "")
}
