% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{pdf_split}
\alias{pdf_split}
\title{Split a pdf file into individual pages}
\usage{
pdf_split(input, output = NULL, password = "")
}
\arguments{
\item{input}{path or url to the input pdf file}

\item{output}{base path of the output file(s)}

\item{password}{string with password to open pdf file}
}
\value{
a character vector with the paths of the split pdf files
}
\description{
Split a pdf file into individual pages
}
\examples{
# extract some pages
pdf_file <- system.file("examples", "sufganiyot.pdf", package = "cpp11qpdf")
fout <- tempfile()
pdf_split(pdf_file, fout, password = "")
}
