% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpdf.R
\name{cpp11qpdf}
\alias{cpp11qpdf}
\alias{pdf_split}
\alias{pdf_length}
\alias{pdf_subset}
\alias{pdf_combine}
\alias{pdf_compress}
\alias{pdf_overlay_stamp}
\alias{pdf_rotate_pages}
\title{PDF Functions}
\usage{
pdf_split(input, output = NULL, password = "")

pdf_length(input, password = "")

pdf_subset(input, pages = 1, output = NULL, password = "")

pdf_combine(input, output = NULL, password = "")

pdf_compress(input, output = NULL, linearize = FALSE, password = "")

pdf_overlay_stamp(input, stamp, output = NULL, password = "")

pdf_rotate_pages(
  input,
  pages,
  angle = 90,
  relative = FALSE,
  output = NULL,
  password = ""
)
}
\arguments{
\item{input}{path or url to the input pdf file}

\item{output}{base path of the output file(s)}

\item{password}{string with password to open pdf file}

\item{pages}{a vector with page numbers to rotate}

\item{linearize}{enable pdf linearization (streamable pdf)}

\item{stamp}{pdf file of which the first page is overlayed into each page of input}

\item{angle}{rotation angle in degrees (positive = clockwise)}

\item{relative}{if \code{TRUE}, pages are rotated relative to their current orientation. If \code{FALSE}, rotation is absolute (0 = portrait, 90 = landscape, rotated 90 degrees clockwise from portrait)}
}
\description{
PDF Functions
}
\examples{
\donttest{
# extract some pages
pdf_file <- file.path(tempdir(), "output.pdf")
pdf_subset('https://cran.r-project.org/doc/manuals/r-release/R-intro.pdf',
  pages = 1:3, output = pdf_file)
pdf_length(pdf_file)
unlink(pdf_file)
}
}
